requirements.txt
Flask
urllib3
requests
python-dotenv
google-ads
cryptography
pymongo
Flask-PyMongo
google-auth
certifi
click
distlib
google-api-core
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
oauth2client
pipenv
pycparser
pyparsing
PyYAML
requests-oauthlib
flask-cors
redis
redis-om
redis-py-cluster
gunicorn
psycopg2-binary
SQLAlchemy
Flask-SQLAlchemy